Simple install (#104)
* Update media-docker.sh

reordering functions

* Update media-docker.sh

self install process, check for existence of .git, download if it doesn't exist, then move to that process

* check if we're a file

* exit function

check that we weren't exiting the installer before cleaning up

* run in subshell

* change dir

* more subshell

* set self as executable

* exit code

* exit code log

* exit code handling

* validations

+++ media-docker.sh
@@ -4,6 +4,27 @@
 # full media server goodness.
 # shellcheck disable=SC2143
 
+#/ Usage: sudo media-docker [OPTIONS]
+#/ Description: media-docker installation and configuration tool.
+#/ 
+#/ To run the GUI installer / configurator, run media-docker without options.
+#/
+#/ Options:
+#/
+#/   -p, --prereq <noremove/test>: Install pre-requisites.
+#/   -a, --apps: View or edit your .apps file
+#/   -e, --env: View or edit your .env file
+#/   -c, --compose <up/down/restart/pull/create>: Rebuild your compose environment from selected options
+#/   -u, --update: Update media-docker
+#/   -P, --prune: Prune the Docker system
+#/   -t, --test: Run tests
+#/   -v, --verbose: Include detailed logging to terminal.
+#/   -x, --debug: Include debug information in terminal.
+#/   -l, --logLevel <n>: Sets log level to defined number `n`. Must be a number.
+#/   -h, --help: Display this help message
+#/
+usage() { grep '^#/' "${SOURCENAME}" | cut -c4- ; exit 0 ; }
+
 set -euo pipefail
 
 ## GLOBALS
@@ -22,6 +43,20 @@ get_source() {
   echo "${SOURCE}"
 }
 
+# script runner
+run_sh() {
+  local DIR="${1:-}"
+  local FILE="${2:-}"
+  shift; shift
+
+  if [[ -f "${DIR}/${FILE}.sh" ]] ; then
+    source "${DIR}/${FILE}.sh"
+    ${FILE} "$@";
+  else
+    log 3 "${DIR}/${FILE}.sh not found."
+  fi
+}
+
 # script info.
 _VERBOSE=3
 readonly ARGS=("$@")
@@ -37,13 +72,13 @@ readonly CONTAINDIR="$BASEDIR/.containers/"
 readonly BACKUPDIR="$BASEDIR/.backups/"
 readonly LOGFILE="$BASEDIR/.logs/media-docker-$(date +%s).log"
 
-# colors
-readonly RED='\033[0;31m'
-readonly YELLOW='\033[1;33m'
-readonly BLUE='\033[34m'
-readonly NOCOL='\033[0m'
-
 log() {
+  # colors
+  local RED='\033[0;31m'
+  local YELLOW='\033[1;33m'
+  local BLUE='\033[34m'
+  local NOCOL='\033[0m'
+
   local -A LOGLVLS=([0]="EMERG" [1]="ALERT" [2]="CRIT" [3]="ERR" [4]="WARN" \
     [5]="NOTICE" [6]="INFO" [7]="DEBUG")
 
@@ -76,59 +111,32 @@ log() {
   fi
 }
 
-# script runner
-run_sh() {
-  local DIR="${1:-}"
-  local FILE="${2:-}"
-  shift; shift
-
-  if [[ -f "${DIR}/${FILE}.sh" ]] ; then
-    source "${DIR}/${FILE}.sh"
-    ${FILE} "$@";
-  else
-    log 3 "${DIR}/${FILE}.sh not found."
-  fi
-}
-
-#/ Usage: sudo media-docker [OPTIONS]
-#/ Description: media-docker installation and configuration tool.
-#/ 
-#/ To run the GUI installer / configurator, run media-docker without options.
-#/
-#/ Options:
-#/
-#/   -p, --prereq <noremove/test>: Install pre-requisites.
-#/   -a, --apps: View or edit your .apps file
-#/   -e, --env: View or edit your .env file
-#/   -c, --compose <up/down/restart/pull/create>: Rebuild your compose environment from selected options
-#/   -u, --update: Update media-docker
-#/   -P, --prune: Prune the Docker system
-#/   -t, --test: Run tests
-#/   -v, --verbose: Include detailed logging to terminal.
-#/   -x, --debug: Include debug information in terminal.
-#/   -l, --logLevel <n>: Sets log level to defined number `n`. Must be a number.
-#/   -h, --help: Display this help message
-#/
-usage() { grep '^#/' "${SOURCENAME}" | cut -c4- ; exit 0 ; }
-
 finish() {
+  local EXIT_CODE="$?"
+  log 7 "Exit code: ${EXIT_CODE}"
   log 6 "Cleaning up after ourselves."
+  if [[ ${EXIT_CODE} -eq 222 ]] ; then
+    log 6 "Installer process, no clean up required."
+  else
+    if [[ $(service --status-all | grep -Fq 'docker') ]] ; then
+      log 6 "Restarting Docker service."
+      sudo service docker restart || true
+    fi
 
-  if [[ $(service --status-all | grep -Fq 'docker') ]] ; then
-    log 6 "Restarting Docker service."
-    sudo service docker restart || true
+    log 6 "Ensuring that we are executable."
+    run_sh "$SCRIPTDIR" "self_symlink" || true
+
+    log 6 "Ensuring that our symlink is set."
+    run_sh "$SCRIPTDIR" "self_symlink" || true
+
+    log 6 "Removing media-docker config temp file."
+    run_sh "$SCRIPTDIR" "self_config_delete" || true
+
+    log 6 "Ensuring testing files are gone."
+    sudo rm -f "${TESTDIR}/.apps-test" || true
+    sudo rm -f "${TESTDIR}/docker-compose-test.yml" || true
+    sudo rm -f "${TESTDIR}/docker-compose.yml" || true
   fi
-
-  log 6 "Ensuring that our symlink is set."
-  run_sh "$SCRIPTDIR" "self_symlink" || true
-
-  log 6 "Removing media-docker config temp file."
-  run_sh "$SCRIPTDIR" "self_config_delete" || true
-
-  log 6 "Ensuring testing files are gone."
-  sudo rm -f "${TESTDIR}/.apps-test" || true
-  sudo rm -f "${TESTDIR}/docker-compose-test.yml" || true
-  sudo rm -f "${TESTDIR}/docker-compose.yml" || true
 }
 trap finish EXIT
 
@@ -136,6 +144,24 @@ trap finish EXIT
 main() {
   local -a PREREQ_CHECKS
   mkdir -p "$(dirname "${LOGFILE}")"
+
+  log 7 "Checking if media-docker is installed."
+  if [[ ! -d ${BASEDIR}/.git ]] ; then
+    _VERBOSE=6
+    log 5 "Downloading media-docker to ${BASEDIR}/media-docker."
+    git clone "https://github.com/joskore/media-docker" \
+      "${BASEDIR}/media-docker" || log 3 "Error occurred when cloning repo."
+
+    log 6 "Cleaning up current file."
+    if [[ -f "$0" ]] ; then
+      sudo rm "$0"
+    fi
+
+    log 6 "Running newly downloaded media-docker."
+    (cd "${BASEDIR}/media-docker/" && \
+      sudo bash "${BASEDIR}/media-docker/media-docker.sh")
+    exit 222
+  fi
 
   log 6 "Checking media-docker requirements."
